refactor: move overrideable behavior onto init service
