+++ testcraft/cli.py
@@ -25,14 +25,14 @@ from craft_application.commands.lifecycle import (
 from craft_providers.actions.snap_installer import Snap
 
 from testcraft.application import TESTCRAFT
-from testcraft.commands import InitCommand, LintCommand, StateCommand
+from testcraft.commands import LintCommand, StateCommand
 from testcraft.services import register_services
 
 
 def register_commands(app: craft_application.Application) -> None:
     """Register extra commands for testcraft."""
     app.add_command_group("Lifecycle", [TestCommand], ordered=True)
-    app.add_command_group("Other", [InitCommand, LintCommand])
+    app.add_command_group("Other", [LintCommand])
     app.add_command_group("State", [StateCommand])
